feat: lift task-details modal to columngroup component
